requirements.txt
asgiref==3.9.2
Django==5.2.7
sqlparse==0.5.3
tzdata==2025.2
gunicorn==21.2.0 # Production WSGI server
psycopg2-binary==2.9.9 # PostgreSQL adapter for production database
python-dotenv==1.0.0 # Load environment variables from .env 

# API Development
djangorestframework==3.14.0 # REST API framework
djangorestframework-simplejwt==5.3.0 # JWT authentication
django-cors-headers==4.3.1 # CORS handling

# Phone number validation
# phonenumbers==8.13.25 # Temporarily disabled
